2024/05/22

+++ springboot/Chapter06-Thymeleaf/src/main/java/edu/pnu/controller/BoardController.java
@@ -1,52 +1,63 @@
 package edu.pnu.controller;
 
+
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.SessionAttributes;
 import org.springframework.web.servlet.ModelAndView;
 
 import edu.pnu.domain.Board;
+import edu.pnu.domain.Member;
 import edu.pnu.service.BoardService;
+
+@SessionAttributes("member")
 @Controller
 public class BoardController {
 	
 	@Autowired
 	private BoardService boardService;
-	
+	@ModelAttribute("member")//초기화
+	public Member setMember() {
+		return new Member();
+	}
 	@GetMapping("/getBoardList")
-	public String getBoardList(Model model, Board board) {
-		System.out.println("getboardlist");
+	public String getBoardList(@ModelAttribute("member") Member member, Model model, Board board) {//세션에 저장되어 있는 member정보를 받아와서 입력
+		if(member.getId()==null) {
+			return "redirect:login";
+		}
 		List<Board> boardList = boardService.getBoardList(board);
-//		for(long i=1L; i<=10L; i++) {
-//			boardList.add(Board.builder()
-//								.seq(i)
-//								.title("게시판 프로그램 테스트")
-//								.writer("도우너")
-//								.content("게시판 프로그램 테스트입니다.")
-//								.createDate(new Date())
-//								.cnt(0L)
-//								.build());
-//		}
+
 		model.addAttribute("boardList", boardList);
 		return "getBoardList";
 	}
 	@GetMapping("/insertBoard")
-	public String insertBoardView() {
+	public String insertBoardView(@ModelAttribute("member") Member member) {
+		if(member.getId()==null) {			
+			return "insertBoard";
+		}
 		return "insertBoard";
 	}
 	
 	@PostMapping("/insertBoard")
-	public String insertBoard(Board board) {
+	public String insertBoard(@ModelAttribute("member") Member member, Board board) {
+		if(member.getId()==null) {
+			return "redirect:getBoardList";
+		}
 		boardService.insertBoard(board);
 		return "redirect:getBoardList";
 	}
 	
 	@GetMapping("/getBoard")
-	public String getBoard(Board board, Model model) {
+	public String getBoard(@ModelAttribute("member") Member member, Board board, Model model) {
+		if(member.getId()==null) {
+			return "redirect:login";
+		}
 		model.addAttribute("board", boardService.getBoard(board));
 		return "getBoard";
 	}
@@ -62,13 +73,19 @@ public class BoardController {
 	}
 	
 	@PostMapping("/updateBoard")
-	public String updateBoard(Board board) {
+	public String updateBoard(@ModelAttribute("member") Member member, Board board) {
+		if(member.getId()==null) {
+			return "redirect:login";
+		}
 		boardService.updateBoard(board);
 		return "redirect:getBoardList";
 	}
 	
 	@GetMapping("/deleteBoard")
-	public String deleteBoard(Board board) {
+	public String deleteBoard(@ModelAttribute("member") Member member, Board board) {
+		if(member.getId()==null) {
+			return "redirect:login";
+		}
 		boardService.deleteBoard(board);
 		return "forward:getBoardList";
 	}
